# openapi.yaml
openapi: 3.0.3
info:
  title: F1 API
  description: OpenAPI specification for the F1 API endpoints.
  version: 1.0.0
servers:
  - url: https://f1api.dev
    description: Production server
...
paths:
...
  /api/{year}/{round}/qualy:
    get:
      tags:
        - races
      summary: "Retrieve the information of the qualy of a race based on the year and the round"
      description: "Retrieve the information of the qualy of a race based on the year and the round."
      operationId: getYearRoundQualy
      parameters:
        - in: path
          name: year
          required: true
          description: "Season year (for example 2024)."
          schema:
            type: integer
        - in: path
          name: round
          required: true
          description: "Race round number within the season."
          schema:
            type: integer
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/timezone'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Resource not found
        '500':
          description: Internal server error
...
components:
  parameters:
    limit:
      in: query
      name: limit
      required: false
      description: Number of records to return. Min 1, max 100.
      schema:
        type: integer
        default: 30
        minimum: 1
        maximum: 100
    offset:
      in: query
      name: offset
      required: false
      description: Number of records to skip before collecting the response.
      schema:
        type: integer
        default: 0
        minimum: 0
        maximum: 10000
...
    timezone:
      in: query
      name: timezone
      required: false
      description: IANA timezone identifier used to convert date/time fields (for example Europe/Madrid).
      schema:
        type: string
        example: Europe/Madrid